+++ CHANGELOG.md
@@ -22,6 +22,9 @@ The version is represented by three digits: a.b.c.
 
 ## Unreleased
 
+MAINTENANCE:
+- added mypy
+
 ## \[0.3.2\] - 2025-02-08
 
 MAINTENANCE: